Revision: cxml--devel--1.0--patch-5 Archive: david@knowledgetools.de--cxml Creator: David Lichteblau <david@knowledgetools.de> Date: Fri Jan  9 16:58:44 CET 2004 Standard-date: 2004-01-09 15:58:44 GMT Modified-files: xml/package.lisp xml/sax-tests/tests.lisp     xml/xml-parse.lisp xml/xml-stream.lisp New-patches: david@knowledgetools.de--cxml/cxml--devel--1.0--patch-5 Summary: (assert (eql initial-speed 1)) in make-xstream Keywords:
MAKE-XSTREAM: xstreams must be created with an INITIAL-SPEED of 1, otherwise
the first buffer will the read with a fixed encoding instead of the encoding
declared in the processing instruction.  Not sure if xstreams are used for
non-XML data, so I've left the argument in, but added an assertion.

PARSE-FILE: works now.

PARSE-STRING remains broken, don't export it